LayoutTests/plugins/mouse-events-fixedpos-expected.txt

CONSOLE MESSAGE: PLUGIN: getFocusEvent
CONSOLE MESSAGE: PLUGIN: mouseDown at (50, 50)
CONSOLE MESSAGE: PLUGIN: mouseUp at (50, 50)
CONSOLE MESSAGE: PLUGIN: mouseDown at (60, 60)
CONSOLE MESSAGE: PLUGIN: mouseUp at (70, 60)

Tests for widget positions being correctly updated after scrolling. rdar://problem/7559069
